requirements.txt
# YOLO Training Interface Dependencies
ultralytics>=8.0.0
rich>=10.0.0
pyyaml>=6.0
pandas>=1.3.0
matplotlib>=3.4.0
numpy>=1.20.0
# Ray is optional for hyperparameter tuning
ray>=2.0.0
# Additional dependencies 
# For Google Colab compatibility
ipython>=7.0.0
